Move back CacheConfig into config package
